Update message handling and improve application performance
The code changes mainly improve the way messages are handled in the MyReactiveMessagingApplication class on both producer and consumer side. Changes to the scheduling method and messaging stream allow for more efficient message handling. In terms of application performance, the Consumer application's batch size for Kafka messages has been increased and the Gift class now contains a timestamp and thread info. The SantaClausService class also got updated to make use of the new Gift class features.